+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `whitelistIngestAddresses` to Workspace
 
+### Fixed
+- Make search config action fields required
+
 ## [5.3.0] - 2023-03-28
 ### Changed
 - Updated README with api v3 example